move base nmapg into utils
